requirements.txt
bcrypt==4.0.1
blinker==1.6.3
cachelib==0.10.2
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
Flask==3.0.0
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
pytest==7.4.2
python-dotenv==1.0.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.0
